To many info, and perhaps to much!
I think we should not have the shared host section for the advance application, like this.
We can have both, a shared host, and several subdomains. We can have git hooks, bare repos... the possibilities are endless. Isn't this a little outside the scope of Yii?

We use shared hosts, and they do allow multiple subdomains, each with their own webroot. So I guess this typically is NOT really the case.